chore: moved to ProtoApi.kt
